requirements.txt
accelerate
codetiming
datasets
dill
hydra-core
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
ray[default]>=2.47.1
torchdata
tensordict<=0.6.2
transformers
wandb
tensorboard
mathruler
math_verify
timm
imageio
loguru
packaging>=20.0
dacite
qwen_vl_utils
scipy
